+++ CHANGELOG.rst
@@ -6,6 +6,14 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a
 Changelog <https://keepachangelog.com/en/1.1.0/>`__.
 
+Unreleased
+----------
+
+Added
+~~~~~
+
+- Add ``is_sensitive`` file field
+
 2.0.0 - 2025-11-26
 ------------------